requirements.txt
# requirements.txt
selenium
requests
python-dotenv